clone-detection/MacOS/requirements.txt
pyppeteer
scikit-image
numpy
requests
beautifulsoup4
scikit-learn
Pillow
fastapi
uvicorn
opencv-python